+++ src/logic/presentation/control/SignUpGraphic.java
@@ -119,7 +119,7 @@ public class SignUpGraphic implements Initializable {
 			Stage stage = (Stage)pane.getScene().getWindow();
 			
 			if(SessionFacade.getSession().getCurrUserType() == Users.SEEKER) {
-				stage.setScene(GraphicHandler.switchScreen(Scenes.ACC_SEEK, new SeekerAccountGraphic(SessionFacade.getSession().getID().intValue())));
+				stage.setScene(GraphicHandler.switchScreen(Scenes.ACC_SEEK, new SeekerAccountGraphic(SessionFacade.getSession().getID())));
 				
 			}else if(SessionFacade.getSession().getCurrUserType() == Users.RECRUITER) {
 				stage.setScene(GraphicHandler.switchScreen(Scenes.ACC_REC, null));
